requirements.txt
# Core AI
google-genai>=0.3.0
python-dotenv>=1.0.0

# Speech
faster-whisper>=0.10.0
pyaudio>=0.2.13
pyttsx3>=2.90
websockets>=12.0

# GUI
PySide6>=6.6.0

# Vision
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Utilities
asyncio>=3.4.3

# MCP (Model Context Protocol) Dependencies
playwright>=1.40.0
pyautogui>=0.9.54
pyperclip>=1.8.2
